Pretty numbers (#99)
* Create pn helper function inside utils folder

The function pn() is being defined a few times in different components. This commit is the first of two commits to solve this issue.

This commit only creates the function inside the utils folder.

Jest is also being introduced for unit tests.

* Use pn from utils folder

This commit removes the multiple definitions of pn found across the project and imports the helper function whenever needed.